requirements.txt
Django==1.5.5
South==0.8.2
distribute==0.6.30
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==18.0
oauthlib==0.6.0
psycopg2==2.5.1
python-linkedin==4.0
python-openid==2.2.5
requests==2.0.1
requests-oauthlib==0.4.0
six==1.4.1
static==0.4
wsgiref==0.1.2
